Implement import & native statements
Import just parses and interprets the file given, not much different functionally to a #include but in the future I'd like to add namespacing to match python implementation.

Native is sort of a forward declaration for the compiler to let it know that there is a function that will be bound at runtime while still allowing it to do static checks. Because it's only bound at runtime there can't be any type inference and it must be declared with the return type (ie. no `def`).