Fix leaky stack & simplify emitted code for returns
Statements should have no effect on the stack, and expressions should push 1 item onto stack. However, this wasn't the case before. Having an expression statement like "1 + 1" or "f()" would leave the value of the expression on the stack without being cleaned up. Now expression statements emit an extra pop to clean up after themselves.

Return values for functions are now passed as a hidden first parameter by the caller, and the arguments are cleaned up by the callee. This leads to the desired effect of having the return value at the top of the stack after a function is called. Void functions have an implicit return value of 0, since all expressions have to add an item to the stack.